refactor: move searchBar to separate folder
